moving logs to another folder
