+++ android/src/main/java/io/parity/signer/screens/settings/logs/LogsScreen.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.screens.logs
+package io.parity.signer.screens.settings.logs
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -12,15 +12,15 @@ import io.parity.signer.components.base.ScreenHeader
 import io.parity.signer.components.panels.BottomBar2
 import io.parity.signer.components.panels.BottomBar2State
 import io.parity.signer.domain.Navigator
-import io.parity.signer.screens.logs.items.LogItem
-import io.parity.signer.screens.logs.items.LogItemDate
-import io.parity.signer.screens.logs.items.LogsListEntryModel
+import io.parity.signer.screens.settings.logs.items.LogItem
+import io.parity.signer.screens.settings.logs.items.LogItemDate
+import io.parity.signer.screens.settings.logs.items.LogsListEntryModel
 import io.parity.signer.uniffi.Action
 
 @Composable
 fun LogsScreen(
-	model: LogsScreenModel,
-	navigator: Navigator,
+    model: LogsScreenModel,
+    navigator: Navigator,
 ) {
 	Column(Modifier.background(MaterialTheme.colors.background)) {
 		ScreenHeader(
