+++ android/src/main/java/io/parity/signer/screens/settings/logs/LogsSubgraph.kt
@@ -1,0 +1,23 @@
+package io.parity.signer.screens.settings.logs
+
+
+//is ScreenData.Log
+//Box(Modifier.statusBarsPadding()) {
+//	LogsScreen(
+//		model = screenData.f.toLogsScreenModel(),
+//		navigator = rootNavigator,
+//	)
+//}
+
+
+//				is ModalData.LogRight ->
+//					BottomSheetWrapperRoot(onClosedAction = {
+//						navigator.backAction()
+//					}) {
+//						LogsMenu(
+//							navigator = sharedViewModel.navigator,
+//						)
+//					}
+
+
+//		is ScreenData.LogDetails -> LogDetails(screenData.f)
