+++ android/src/main/java/io/parity/signer/screens/settings/logs/logdetails/LogDetails.kt
@@ -1,10 +1,10 @@
-package io.parity.signer.screens.logs.logdetails
+package io.parity.signer.screens.settings.logs.logdetails
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import io.parity.signer.screens.logs.logdetails.HistoryCardExtended
+import io.parity.signer.screens.settings.logs.logdetails.HistoryCardExtended
 import io.parity.signer.uniffi.MLogDetails
 
 @Composable
