+++ android/src/main/java/io/parity/signer/ui/rustnavigationselectors/ScreensSelectorsLegacy.kt
@@ -11,7 +11,7 @@ import io.parity.signer.domain.*
 import io.parity.signer.domain.storage.addSeed
 import io.parity.signer.domain.storage.signSufficientCrypto
 import io.parity.signer.screens.*
-import io.parity.signer.screens.logs.logdetails.LogDetails
+import io.parity.signer.screens.settings.logs.logdetails.LogDetails
 import io.parity.signer.screens.settings.VerifierScreen
 import io.parity.signer.ui.theme.SignerOldTheme
 import io.parity.signer.uniffi.Action
@@ -49,7 +49,7 @@ fun ScreenSelector(
 		}
 		is ScreenData.Keys -> {} //migrated to new selector
 		is ScreenData.Log -> {} //migrated to new selector
-		is ScreenData.LogDetails -> LogDetails(screenData.f)
+		is ScreenData.LogDetails -> {} // moved to settings flow, not part of global state machine now
 		is ScreenData.ManageNetworks -> {} //migrated to new selector
 		is ScreenData.NNetworkDetails -> {} // migrated to new selector
 		is ScreenData.NewSeed -> {} // new selector
